Compiler flags (#107)
* added -Werror -Wall

* fixed brure force init order

* removed unused variable

* size_t comparison

* bf_batch_iterator init order

* invalid id is uint max

* hnsw batch iterator init format

* compare levels as int

* levels are size_t

* wip

* fixed hnsw and tests

* fixed LLVM warnings

* fixed valgrind build

* undo change

+++ tests/unit/test_utils.h
@@ -4,12 +4,12 @@
 #include "VecSim/vec_sim.h"
 
 void runTopKSearchTest(VecSimIndex *index, const void *query, size_t k,
-                       std::function<void(int, float, int)> ResCB,
+                       std::function<void(size_t, float, size_t)> ResCB,
                        VecSimQueryParams *params = nullptr,
                        VecSimQueryResult_Order order = BY_SCORE);
 
 void runBatchIteratorSearchTest(VecSimBatchIterator *batch_iterator, size_t n_res,
-                                std::function<void(int, float, int)> ResCB,
+                                std::function<void(size_t, float, size_t)> ResCB,
                                 VecSimQueryResult_Order order = BY_SCORE,
                                 size_t expected_n_res = -1);
